requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
flake8==7.1.1
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytz==2024.2
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
